move WorkerState to wire package
